fix :: hook 분리
